test(dashboard): #1556 route 1 — lift the #1487 classifier out and pin annotation coverage by module
#1556 annotates the package's blind failure returns slice by slice. Nothing pinned the result:
`_SIGNED` in `test_collapsed_return_channels.py` is the only set law 1 enforces, and slice 1 added
nothing to it, so a dropped `| None` reverts in silence — the function returns to the blind set,
the collapse count does not move, and the #1487 laws have nothing to say because they only judge a
function that declared a contract. Measured, not argued: reverting `get_stats`'s annotation in
`client/xvb_client.py` leaves `test_collapsed_return_channels.py` GREEN at 18 passed, with the
blind count moving 52 -> 53 in output nothing asserts on.

The pins cannot go in that file: it sat at its recorded 563-line ceiling and `lint-file-budget.sh`
rejects any raise. A sibling test module is not importable by name either — `--import-mode=importlib`
with no `__init__.py` makes `from test_collapsed_return_channels import classify` a
ModuleNotFoundError, checked by running it. So the shared machinery moves to a non-test module.

- `tests/service/annotation_gate.py` (new, 241 lines): the eight classifier functions, moved. The
  186-line block is byte-identical to its origin apart from one line — `_classify` is renamed
  `classify`, since it now crosses a module boundary. The package walk and its empty-scan guard
  become `classify_package()`, so a second caller cannot inherit the walk without the guard.
- `tests/service/test_collapsed_return_channels.py`: 563 -> 353 lines. It keeps the argument, both
  laws, `_SIGNED` and all 18 tests, unchanged — the 18 collected IDs are identical before and after.
  Its budget row is DELETED rather than lowered: a file at or under the 400 target takes no row, and
  a needless entry fails the gate as loudly as a missing one.
- `tests/service/test_annotation_coverage.py` (new, 163 lines): the ratchet, pinning BY MODULE —
  "no failure return in this module is unannotated" — which a rename cannot disturb and a new
  function cannot slip past. Two modules qualify at this tip, measured over live source:
  `client/xvb_client.py` (slice 1) and `service/worker_config_store.py`, whose
  `note_worker_revision` was signed-but-unpinned already at the base of this branch.

This pin is not the baseline #1487 refused. That ruling was about the collapsed instances: a `<= 11`
bound would certify unread defects as intentional. Here every function in a pinned module was read
as part of the slice that annotated it, and what is asserted is that the reading happened — never
that the verdict is acceptable. A pinned module may still hold a reported collapse and this file
stays green while it does. What it refuses is the module going quiet again.

A nine-mutation battery over the new module found a real defect in it, which is fixed here rather
than shipped: the vacuity control asserted `_rows_under(module, <hand-built empty dict>) == []`,
which no mutation could kill and which its own docstring overstated — an absence satisfiable by a
predicate that returns nothing for everything, the shape a broken walk actually has. It now runs
against the REAL package and asserts both directions, absent module and present sibling, and two
mutations kill it. After the fix every one of the 8 tests is killed by at least one mutation and
every pair is separated by one, so none is a duplicate.

Docs: `docs/dashboard.md` and `docs/dev/repo-map.md` — the return-channel rule now lives in two test
files over one shared classifier, and `dashboard/tests/` now holds a `.py` that is not a test.
