+++ dashboard/tests/service/test_collapsed_return_channels.py
@@ -75,22 +75,9 @@ is what drives the `payout_confirmed` alert. The alert that does not fire is the
 symptom.
 """
 
-import ast
-import pathlib
-
 import pytest
 
-_PACKAGE = pathlib.Path(__file__).resolve().parents[2] / "mining_dashboard"
-
-# The collapsed values an error path can hide inside a success type. `False` is deliberately NOT
-# here: it was measured at 5 further instances, all of them outbound senders (`_post`/`ping`/
-# `send`) where False-on-failure and False-elsewhere mean the same thing to a caller. Including it
-# would add five probable false positives to a gate whose whole value is that its findings are real.
-_EMPTY = {"[]", "{}", "0"}
-
-# The DB handle. A guard testing THIS is a failure route; a guard testing a domain value is an
-# early return. That distinction is what separates 32 in-scope sites from 139 deliberately excluded.
-_HANDLE = "_conn"
+from tests.service.annotation_gate import classify, classify_package
 
 # Measured, read at source, and cleared — the only set this file ratchets. Each declares `T | None`
 # and returns `None` from every failure path: the three-valued contract #1409 established,
@@ -105,207 +92,10 @@ _SIGNED = {
 }
 
 
-def _collapsed_literal(value: ast.AST | None) -> str | None:
-    """The collapsed value a ``return`` yields, or None if it is not one of the shapes we track.
-
-    An empty list/dict written as a literal, a bare `return`, `None`, and integer `0`. `False` is
-    reported so the classifier can SEE it, and excluded from `_EMPTY` so it cannot be flagged — the
-    two are different decisions and collapsing them here would be this file's own defect. `0` is
-    guarded against `bool`, because `False == 0` is true in Python and an unguarded check would
-    silently reclassify every boolean return as a numeric collapse.
-    """
-    if value is None:
-        return "<bare>"
-    if isinstance(value, ast.List) and not value.elts:
-        return "[]"
-    if isinstance(value, ast.Dict) and not value.keys:
-        return "{}"
-    if isinstance(value, ast.Constant):
-        if value.value is None:
-            return "None"
-        if value.value is False:
-            return "False"
-        if value.value == 0 and isinstance(value.value, int) and not isinstance(value.value, bool):
-            return "0"
-    return None
-
-
-def _own_nodes(function: ast.AST) -> list[ast.AST]:
-    """Every node belonging to THIS function, never to a def nested inside it.
-
-    A nested def is its own function with its own contract: folding its returns into the parent
-    both double-counts and misattributes. `web/server.py:_log_filters` and the `_num` closure
-    inside it are the real case — a naive walk reports them as one function and gets the answer
-    wrong in both directions at once.
-    """
-    collected: list[ast.AST] = []
-
-    def walk(node: ast.AST) -> None:
-        for child in ast.iter_child_nodes(node):
-            if isinstance(child, (ast.FunctionDef, ast.AsyncFunctionDef)):
-                continue
-            collected.append(child)
-            walk(child)
-
-    walk(function)
-    return collected
-
-
-def _union_members(annotation: ast.AST) -> list[ast.AST]:
-    """The TOP-LEVEL alternatives of a return annotation, flattened.
-
-    Only the top level, which is the entire point. `None` nested inside a parameter — the return
-    slot of a `Callable[[], None]`, the value type of a `dict[str, None]` — says nothing about
-    whether THIS function can return None, and a walk of the whole annotation tree cannot tell the
-    two apart. Nothing in the package wears that shape today; this is written so the gate does not
-    start misreading the day something does, because the failure would be a false accusation
-    against a function that never collapsed anything.
-    """
-    if isinstance(annotation, ast.BinOp) and isinstance(annotation.op, ast.BitOr):
-        return _union_members(annotation.left) + _union_members(annotation.right)
-    if isinstance(annotation, ast.Subscript):
-        base = annotation.value
-        name = base.id if isinstance(base, ast.Name) else getattr(base, "attr", "")
-        if name == "Optional":
-            # Both halves: `Optional[T]` is `T | None`, and returning only the None would make the
-            # annotation indistinguishable from a bare `-> None` procedure.
-            return _union_members(annotation.slice) + [ast.Constant(None)]
-        if name == "Union":
-            inner = annotation.slice
-            elements = inner.elts if isinstance(inner, ast.Tuple) else [inner]
-            return [member for element in elements for member in _union_members(element)]
-    return [annotation]
-
-
-def _returns_nothing(annotation: ast.AST | None) -> bool:
-    """Is this declared `-> None` — a procedure with no return value at all?
-
-    Distinct from `T | None`, which is a value type carrying an out-of-band failure marker. The
-    procedure has no answer channel for an error to share, so it is neither signed nor collapsed;
-    it is simply not what this file is about.
-    """
-    members = _union_members(annotation) if annotation is not None else []
-    return len(members) == 1 and isinstance(members[0], ast.Constant) and members[0].value is None
-
-
-def _admits_none(annotation: ast.AST | None) -> bool | None:
-    """Does the declared return type carry an out-of-band failure marker?
-
-    None (the return value, not the type) means UNANNOTATED — a third answer, not a false one. A
-    function that declares no type declares no contract, so it can be neither signed nor in breach,
-    and the reporting test counts it as blind rather than clean. Conflating "no annotation" with
-    "does not admit None" would flag the whole unannotated layer as collapsed and make the gate's
-    findings worthless.
-    """
-    if annotation is None:
-        return None
-    return any(
-        isinstance(member, ast.Constant) and member.value is None
-        for member in _union_members(annotation)
-    )
-
-
-def _is_handle_guard(node: ast.AST) -> bool:
-    """``if not self._conn:`` or ``if self._conn is None:`` — a resource handle, not a domain value.
-
-    Both spellings, because the same failure written the other way is the same failure and a gate
-    that saw only one would be muted by a reformat.
-    """
-    if not isinstance(node, ast.If):
-        return False
-    test = node.test
-    if isinstance(test, ast.UnaryOp) and isinstance(test.op, ast.Not):
-        test = test.operand
-    if isinstance(test, ast.Compare) and len(test.ops) == 1 and isinstance(test.ops[0], ast.Is):
-        test = test.left
-    return (
-        isinstance(test, ast.Attribute)
-        and test.attr == _HANDLE
-        and isinstance(test.value, ast.Name)
-        and test.value.id == "self"
-    )
-
-
-def _failure_returns(function: ast.AST) -> list[tuple[str, str, int]]:
-    """``(shape, literal, lineno)`` for every failure return in one function.
-
-    Both doors: a return lexically inside an `except` handler, and a return that is the entire body
-    of a handle guard. A guard return that is itself inside a handler is attributed to the handler
-    once rather than counted twice.
-    """
-    nodes = _own_nodes(function)
-    handler_lines: set[int] = set()
-    found: list[tuple[str, str, int]] = []
-    for node in nodes:
-        if not isinstance(node, ast.ExceptHandler):
-            continue
-        for sub in ast.walk(node):
-            handler_lines.add(getattr(sub, "lineno", -1))
-            if isinstance(sub, ast.Return) and (literal := _collapsed_literal(sub.value)):
-                found.append(("except", literal, sub.lineno))
-    for node in nodes:
-        if not (_is_handle_guard(node) and len(node.body) == 1):
-            continue
-        statement = node.body[0]
-        if not isinstance(statement, ast.Return) or statement.lineno in handler_lines:
-            continue
-        if literal := _collapsed_literal(statement.value):
-            found.append(("guard", literal, statement.lineno))
-    return found
-
-
-def _classify(source: str, where: str) -> dict[str, list]:
-    """Sort every failure return in one module into the four verdicts.
-
-    ``signed`` — declares an out-of-band marker and honours it.
-    ``half_fix`` — declares one and returns an empty container anyway (law 2's quarry).
-    ``collapse`` — declares a type the failure value fits, so no caller can tell them apart.
-    ``blind`` — unannotated, and therefore outside what this mechanism can judge.
-    """
-    verdicts: dict[str, list] = {"signed": [], "half_fix": [], "collapse": [], "blind": []}
-    for function in ast.walk(ast.parse(source)):
-        if not isinstance(function, (ast.FunctionDef, ast.AsyncFunctionDef)):
-            continue
-        failures = _failure_returns(function)
-        if not failures:
-            continue
-        name = f"{where}:{function.name}"
-        # A `-> None` procedure is out of scope entirely, not "signed". It has no success value for
-        # a failure to hide inside, so the defect this file describes cannot occur in one. Counting
-        # its bare `return` as a signed contract would inflate the cleared set with functions that
-        # never made the promise — three of them here — and make the category mean two things.
-        if _returns_nothing(function.returns):
-            continue
-        admits = _admits_none(function.returns)
-        values = {literal for _, literal, _ in failures}
-        # A bare `return` and `return None` are the same value; only the spelling differs. Holding
-        # them apart here would red the gate on a signed function the day someone dropped the
-        # word `None`, which is a false alarm against code that honours the contract exactly.
-        values = {"None" if value == "<bare>" else value for value in values}
-        if admits is None:
-            verdicts["blind"].append(name)
-        elif admits and values <= {"None"}:
-            verdicts["signed"].append(name)
-        elif admits and values & _EMPTY:
-            verdicts["half_fix"].append((name, sorted(values & _EMPTY)))
-        elif not admits and values & _EMPTY:
-            verdicts["collapse"].append((name, sorted(values & _EMPTY)))
-    return verdicts
-
-
 @pytest.fixture(scope="module")
 def package() -> dict[str, list]:
     """Every failure return in `mining_dashboard`, classified, in one pass over the real package."""
-    total: dict[str, list] = {"signed": [], "half_fix": [], "collapse": [], "blind": []}
-    files = sorted(_PACKAGE.rglob("*.py"))
-    # An empty scan is not a clean scan. Every assertion below is a statement about a set, and a
-    # set that is empty because the walk found no files satisfies all of them for the wrong reason.
-    assert len(files) > 50, "enumeration guard: the package walk found almost nothing"
-    for path in files:
-        where = str(path.relative_to(_PACKAGE))
-        for verdict, rows in _classify(path.read_text(), where).items():
-            total[verdict].extend(rows)
-    return total
+    return classify_package()
 
 
 class TestTheSignedContractHolds:
@@ -422,14 +212,14 @@ class Store:
     def test_it_flags_an_error_returning_a_value_its_success_type_admits(self):
         """POSITIVE CONTROL for the `except` shape — the exact #1409 collapse."""
         assert self._COLLAPSE.count("-> list[dict]:") == 1  # arming readback: no `| None` present
-        assert _classify(self._COLLAPSE, "m.py")["collapse"] == [("m.py:get_rows", ["[]"])]
+        assert classify(self._COLLAPSE, "m.py")["collapse"] == [("m.py:get_rows", ["[]"])]
 
     def test_it_clears_the_same_function_once_the_failure_is_out_of_band(self):
         """NEGATIVE CONTROL, and the one that proves the classifier reads the ANNOTATION rather
         than the literal. Same function, same call, same shape — only the contract widened and the
         error value moved out of the success type. A checker keying on `return []` alone would be
         unable to tell these two apart, and would flag #1409's own fix as the defect it fixed."""
-        verdicts = _classify(self._SIGNED_SHAPE, "m.py")
+        verdicts = classify(self._SIGNED_SHAPE, "m.py")
         assert verdicts["signed"] == ["m.py:get_rows"]
         assert verdicts["collapse"] == []
 
@@ -437,14 +227,14 @@ class Store:
         """POSITIVE CONTROL for LAW 2, whose product on real source is an ABSENCE. An empty result
         there means nothing until the probe is shown able to return a non-empty one. This is the
         half-done fix exactly: handler corrected, handle guard left behind."""
-        assert _classify(self._HALF_FIX, "m.py")["half_fix"] == [("m.py:get_rows", ["[]"])]
+        assert classify(self._HALF_FIX, "m.py")["half_fix"] == [("m.py:get_rows", ["[]"])]
 
     def test_it_sees_the_handle_guard_with_no_exception_anywhere(self):
         """POSITIVE CONTROL for the silent half of #1409. Nothing in this function raises, so a
         rule anchored on `except` reports it clean — which is how the route a reader could not see
         stayed invisible in the first place."""
         assert "except" not in self._HANDLE_GUARD
-        assert _classify(self._HANDLE_GUARD, "m.py")["collapse"] == [("m.py:get_rows", ["[]"])]
+        assert classify(self._HANDLE_GUARD, "m.py")["collapse"] == [("m.py:get_rows", ["[]"])]
 
     def test_it_leaves_a_domain_guard_alone(self):
         """NEGATIVE CONTROL, and the sibling that pins the narrowness. Byte-for-byte the same
@@ -453,7 +243,7 @@ class Store:
         every positive control above would still pass — while the gate started firing on the 139
         domain-value guards this rule deliberately leaves alone."""
         assert "        if not rows:\n            return []" in self._DOMAIN_GUARD
-        assert _classify(self._DOMAIN_GUARD, "m.py") == {
+        assert classify(self._DOMAIN_GUARD, "m.py") == {
             "signed": [],
             "half_fix": [],
             "collapse": [],
@@ -463,7 +253,7 @@ class Store:
     def test_an_unannotated_function_is_blind_rather_than_clean(self):
         """The third answer. This is the same body as the positive control with its annotation
         removed, and calling it 'clean' would let any collapse escape by deleting a type hint."""
-        verdicts = _classify(self._UNANNOTATED, "m.py")
+        verdicts = classify(self._UNANNOTATED, "m.py")
         assert verdicts["blind"] == ["m.py:get_rows"]
         assert verdicts["collapse"] == []
 
@@ -480,7 +270,7 @@ class Store:
             "        except Exception:\n"
             "            return False\n"
         )
-        assert _classify(seeded, "m.py")["collapse"] == []
+        assert classify(seeded, "m.py")["collapse"] == []
 
     def test_the_walk_finds_the_collapses_that_really_are_in_the_package(self, package):
         """Corroboration on REAL source rather than seeded strings. `get_payouts` returns `[]` from
@@ -502,7 +292,7 @@ class Store:
                 "        try:\n            return build()\n"
                 "        except Exception:\n            return {}\n"
             )
-            verdicts = _classify(seeded, "m.py")
+            verdicts = classify(seeded, "m.py")
             assert verdicts["signed"] == [], annotation
             assert verdicts["collapse"] == [("m.py:get", ["{}"])], annotation
 
@@ -516,7 +306,7 @@ class Store:
                 "        try:\n            return [1]\n"
                 "        except Exception:\n            return None\n"
             )
-            assert _classify(seeded, "m.py")["signed"] == ["m.py:get"], annotation
+            assert classify(seeded, "m.py")["signed"] == ["m.py:get"], annotation
 
     def test_a_bare_return_honours_the_contract_the_same_as_return_none(self):
         """`return` and `return None` are one value in two spellings. Holding them apart would red
@@ -527,7 +317,7 @@ class Store:
             "        if not self._conn:\n            return\n"
             "        return [1]\n"
         )
-        assert _classify(seeded, "m.py")["signed"] == ["m.py:get"]
+        assert classify(seeded, "m.py")["signed"] == ["m.py:get"]
 
     def test_a_procedure_returning_nothing_is_not_counted_as_a_signed_contract(self):
         """NEGATIVE CONTROL. `reconcile_worker_config_status` and two siblings are `-> None`
@@ -540,7 +330,7 @@ class Store:
             "        if not self._conn:\n            return\n"
             "        self._conn.execute('x')\n"
         )
-        assert _classify(seeded, "m.py") == {
+        assert classify(seeded, "m.py") == {
             "signed": [],
             "half_fix": [],
             "collapse": [],
@@ -560,4 +350,4 @@ class Store:
             "            return []\n"
             "    return 'x'\n"
         )
-        assert _classify(seeded, "m.py")["collapse"] == [("m.py:inner", ["[]"])]
+        assert classify(seeded, "m.py")["collapse"] == [("m.py:inner", ["[]"])]
